Transition from ex-GORM structs to sqlc structs (3/5)
Replace old used-to-be-GORM datastructures (#104305) with sqlc-generated
structs. This also makes it possible to use more specific structs that
are more taylored to the specific queries, increasing efficiency.

This commit deals with worker tags (on both workers and jobs).

Functional changes are kept to a minimum, as the API still serves the
same data.

Because this work covers so much of Flamenco's code, it's been split up
into different commits. Each commit brings Flamenco to a state where it
compiles and unit tests pass. Only the result of the final commit has
actually been tested properly.

Ref: #104343

+++ internal/manager/persistence/jobs_test.go
@@ -97,7 +97,7 @@ func TestStoreAuthoredJobWithWorkerTag(t *testing.T) {
 	require.NotNil(t, fetchedJob)
 
 	require.NotNil(t, fetchedJob.WorkerTagID)
-	assert.Equal(t, *fetchedJob.WorkerTagID, workerTag.ID)
+	assert.Equal(t, int64(*fetchedJob.WorkerTagID), workerTag.ID)
 
 	require.NotNil(t, fetchedJob.WorkerTag)
 	assert.Equal(t, fetchedJob.WorkerTag.Name, workerTag.Name)
